refactor scheduled messages api
